+++ src/commands/kube/deployment/update.ts
@@ -1,7 +1,8 @@
-import { Command, flags } from '@oclif/command';
+import { flags } from '@oclif/command';
 import { executeCommand } from '../../../modules/shared/helpers/execute-command';
+import { BaseCommand } from '../../../modules/shared/base-commands/base-command';
 
-export default class KubeDeploymentUpdate extends Command {
+export default class KubeDeploymentUpdate extends BaseCommand {
     static description = 'Trigger an update on a deployment';
 
     static aliases = ['kdu'];
@@ -26,8 +27,6 @@ export default class KubeDeploymentUpdate extends Command {
         this.log(`Updating deployment ${args.deployment}`);
         await executeCommand(
             `kubectl patch deployment "${args.deployment}" --namespace=${args.namespace} -p "{\\"spec\\":{\\"template\\":{\\"metadata\\":{\\"labels\\":{\\"date\\":\\"\`date +'%s'\`\\"}}}}}"`,
-            this.log,
-            this.error,
         );
     }
 }