+++ src/modules/shared/base-commands/build-image-workflow.base-command.ts
@@ -1,10 +1,10 @@
+import { BaseCommand } from './base-command';
 import { buildDockerImage } from '../helpers/docker/build-docker-image';
 import { pushDockerImage } from '../helpers/docker/push-docker-image';
-import { getHashValue } from '../helpers/entro-hash/get-hash-value';
-import { BaseCommand } from './base-command';
-import { setHashValue } from '../helpers/entro-hash/set-hash-value';
 import { generateDockerImageName } from '../helpers/docker/docker-image-name-builder';
 import { pullDockerImage } from '../helpers/docker/pull-docker-image';
+import { container } from 'tsyringe';
+import { BuildArtefactService, BuildTrigger } from '../services/build-artefact.service';
 
 export interface BuildFromHashFlags {
     // The name of the image that is pushed to the registry
@@ -33,120 +33,45 @@ export abstract class BuildImageWorkflowBaseCommand extends BaseCommand {
      * @param flags The flags required during the build process
      */
     async buildFromHash(hash: string, directory: string, flags: BuildFromHashFlags) {
-        const baseBuild = await this.checkBaseHashAndBuild(hash, directory, flags);
-        await this.checkAndPushHashWithImageName(hash, directory, flags, baseBuild.builtLocally);
-    }
-
-    /**
-     * Checks the main hash and only builds the image if the base hash doesn't
-     * exit
-     * @param hash The base hash that we're checking
-     * @param directory The directory where Dockerfile is
-     * @param flags The flags for the build process
-     * @private
-     */
-    private async checkBaseHashAndBuild(hash: string, directory: string, flags: BuildFromHashFlags) {
-        const exists = await getHashValue(hash, await this.getConfig());
+        const buildArtifactService = container.resolve(BuildArtefactService);
+        const exists = await this.dockerImageExists(hash, flags);
         if (exists) {
-            // If it already exists, there is no need to build it again, we
-            // don't pull the image unless we need it for the tags
-            this.log(`The image already exists, there is no need to build it again!`);
-            return {
-                builtLocally: false,
-                localImageName: generateDockerImageName(flags.imageName, 'local-build'),
-            };
-        }
-
-        // Build the image locally
-        const localImageName = await buildDockerImage(
-            directory,
-            flags.imageName,
-            this.log,
-            this.warn,
-            flags.dockerFileName,
-            flags.registry,
-            flags.dryRun,
-        );
-
-        // Push the image to the registry
-        await pushDockerImage(
-            localImageName,
-            this.getTags(hash, flags),
-            flags.imageName,
-            this.log,
-            this.warn,
-            flags.dryRun,
-            flags.registry,
-        );
-
-        // Set the hash now that it has been built
-        await setHashValue(
-            hash,
-            {
-                hash: true,
+            const versionExists = await this.dockerImageExists(flags.tag, flags);
+            if (versionExists) {
+                await buildArtifactService.addBuild(flags.imageName, BuildTrigger.exists);
+            } else {
+                await buildArtifactService.addBuild(flags.imageName, BuildTrigger.hashExists);
+            }
+            // If it exists, we don't need to build it again but we should push
+            //  it to its new tag.
+            const existingImageName = generateDockerImageName(flags.imageName, hash, flags.registry);
+            await pushDockerImage({
+                localImageName: existingImageName,
                 imageName: flags.imageName,
-            },
-            await this.getConfig(),
-        );
-
-        return {
-            builtLocally: true,
-            localImageName,
-        };
-    }
-
-    /**
-     * Sometimes, we need to push the base to all image names even if the base
-     * hash has already been built
-     * @param hash The hash that we're looking for
-     * @param directory The directory with the Dockerfile
-     * @param flags The flags for the build process
-     * @param baseBuiltLocally Whether or not the base was already built in
-     * this process or if it needs to be pulled
-     * @private
-     */
-    private async checkAndPushHashWithImageName(
-        hash: string,
-        directory: string,
-        flags: BuildFromHashFlags,
-        baseBuiltLocally: boolean,
-    ) {
-        const imageHash = `${flags.imageName}-${hash}`;
-        const exists = await getHashValue(imageHash, await this.getConfig());
-        if (exists) {
-            // There's nothing to do from this point
-            this.log(`The image ${flags.imageName}, there is no need to build it again!`);
+                tags: [flags.tag],
+                registry: flags.registry,
+                dryRun: flags.dryRun,
+            });
             return;
         }
-        if (!baseBuiltLocally) {
-            try {
-                // This was built before, we just need to pull it so that we can
-                // push it under a different image name
-                this.log(`Pulling the base image from the registry`);
-                await pullDockerImage(hash, flags.imageName, this.log, this.warn, flags.registry, flags.dryRun);
-            } catch (e) {
-                this.log(`Failed to pull the image so we'll just build it!`);
-            }
-        }
-        const baseImageName = generateDockerImageName(flags.imageName, hash, flags.registry);
-        await pushDockerImage(
-            baseImageName,
-            this.getTags(hash, flags),
-            flags.imageName,
-            this.log,
-            this.warn,
-            flags.dryRun,
-            flags.registry,
-        );
-        await setHashValue(
-            imageHash,
-            {
-                hash: true,
-                imageName: flags.imageName,
-            },
-            await this.getConfig(),
-        );
-        this.log(`Pushed the base image to ${flags.imageName}`);
+        // Build the image locally
+        const localImageName = await buildDockerImage({
+            directory,
+            imageName: flags.imageName,
+            dryRun: flags.dryRun,
+            dockerFileName: flags.dockerFileName,
+        });
+
+        // Push the image to the registry
+        await pushDockerImage({
+            localImageName,
+            tags: this.getTags(hash, flags),
+            imageName: flags.imageName,
+            registry: flags.registry,
+            dryRun: flags.dryRun,
+        });
+
+        await buildArtifactService.addBuild(flags.imageName, BuildTrigger.new);
     }
 
     /**
@@ -161,5 +86,24 @@ export abstract class BuildImageWorkflowBaseCommand extends BaseCommand {
             tags.push(flags.tag);
         }
         return tags;
+    }
+
+    /**
+     * Check if the docker image has been built before
+     * @param hash The hash that should be attached to the image
+     * @param flags The flags for the build
+     * @private
+     */
+    private async dockerImageExists(hash: string, flags: BuildFromHashFlags) {
+        const dockerImageName = generateDockerImageName(flags.imageName, hash, flags.registry);
+        this.log(`Checking for ${dockerImageName}`);
+        try {
+            await pullDockerImage(hash, flags.imageName, flags.registry, flags.dryRun);
+            this.log(`The docker image ${dockerImageName} has already been built before!`);
+            return true;
+        } catch (e) {
+            this.log(`The docker image ${dockerImageName} has NOT been built before!`);
+            return false;
+        }
     }
 }