+++ src/modules/shared/helpers/docker/docker-image-exists.ts
@@ -1,16 +1,9 @@
 import { executeCommand } from '../execute-command';
 import { generateDockerImageName } from './docker-image-name-builder';
-import { Logger } from '../../interfaces/logger';
 
-export async function dockerImageExists(
-    imageName: string,
-    hash: string,
-    log: Logger,
-    error: Logger,
-    registry?: string,
-): Promise<boolean> {
+export async function dockerImageExists(imageName: string, hash: string, registry?: string): Promise<boolean> {
     try {
-        await executeCommand(`docker pull ${generateDockerImageName(imageName, hash, registry)}`, log, error, false);
+        await executeCommand(`docker pull ${generateDockerImageName(imageName, hash, registry)}`, false);
         return true;
     } catch (e) {
         return false;