+++ src/modules/shared/helpers/docker/push-docker-image.ts
@@ -1,27 +1,27 @@
 import { generateDockerImageName } from './docker-image-name-builder';
 import { executeCommand } from '../execute-command';
-import { Logger } from '../../interfaces/logger';
 import cli from 'cli-ux';
+import { container } from 'tsyringe';
+import { Logger } from '../../services/logger';
 
-export async function pushDockerImage(
-    localImageName: string,
-    tags: string[],
-    imageName: string,
-    log: Logger,
-    error: Logger,
-    dryRun: boolean,
-    registry?: string,
-) {
+export interface PushDockerImageParams {
+    localImageName: string;
+    tags: string[];
+    imageName: string;
+    dryRun: boolean;
+    registry: string;
+}
+
+export async function pushDockerImage({ localImageName, tags, imageName, dryRun, registry }: PushDockerImageParams) {
+    const logger = container.resolve(Logger);
     for (const tag of tags) {
         const dockerImageName = generateDockerImageName(imageName, tag, registry);
         cli.action.start(`Pushing ${dockerImageName}...`);
         await executeCommand(
             `docker tag ${localImageName} ${dockerImageName} && docker push ${dockerImageName}`,
-            log,
-            error,
             dryRun,
         );
         cli.action.stop();
-        log(`Successfully pushed ${dockerImageName}`);
+        logger.log(`Successfully pushed ${dockerImageName}`);
     }
 }