+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [2.0.2](https://github.com/entrostat/entro-ci/compare/v2.0.1...v2.0.2) (2021-02-06)
+
+
+### Bug Fixes
+
+* **reflect-metadata:** Added the import for reflect-metadata to ensure that the CLI builds and runs correctly ([3483c95](https://github.com/entrostat/entro-ci/commit/3483c958bf3801901c0fb97f3b2da161a223b6be))
+
+### [2.0.1](https://github.com/entrostat/entro-ci/compare/v2.0.0...v2.0.1) (2021-02-06)
+
+
+### Bug Fixes
+
+* **logger:** Use the logger instead of passing parameters ([6de012b](https://github.com/entrostat/entro-ci/commit/6de012b0f5f7f47328bdb5191132cbfdb86cb6ce))
+
+## [2.0.0](https://github.com/entrostat/entro-ci/compare/v1.5.7...v2.0.0) (2021-02-06)
+
+
+### Features
+
+* **docker-build:** Completely changed the build process to only use the docker registry to determine if something has been built and added the ability to track whether or not the build happened now. ([2c2feb9](https://github.com/entrostat/entro-ci/commit/2c2feb95221b451a6330f1ff21cc2cce72108af5)), closes [#1](https://github.com/entrostat/entro-ci/issues/1)
+
 ### [1.5.7](https://github.com/entrostat/entro-ci/compare/v1.5.6...v1.5.7) (2021-01-21)